streamline and reduce memory footprint
created function to handle computing apen and sampen for multiple values of m on the same data sequence. function called compute_apen_and_sampen. Moved the list of mvals to a global default variable called MVALS, just like SIG2V.

Renamed cgr_renyi() to compute_cgr_renyi() to make it a little more clear and also follow the format of compute_apen_and_sampen().

added lines to save the sample set for each nobs within the for loop so that only one set of samples is in memory at a time. the variable "simulated" is now a dict where each key is the sampname and each value is the file path of the saved sample data.  data are saved as compressed numpy binary files.